dailyfresh/requirements.txt
amqp==2.4.2
billiard==3.5.0.5
celery==4.2.1
certifi==2019.3.9
chardet==3.0.4
Django==2.1.7
django-haystack==2.8.1
django-redis==4.10.0
django-tinymce==2.8.0
dnspython==1.16.0
eventlet==0.24.1
fdfs-client-py==1.2.6
greenlet==0.4.15
idna==2.8
itsdangerous==1.1.0
jieba==0.39
kombu==4.4.0
monotonic==1.5
mutagen==1.42.0
Pillow==5.4.1
pycryptodomex==3.7.2
PyMySQL==0.9.3
python-alipay-sdk==1.10.0
pytz==2018.9
redis==3.2.0
requests==2.21.0
six==1.12.0
urllib3==1.24.1
vine==1.2.0
Whoosh==2.7.4
